Project structure updated - Sources placed in it's own directory - Displays added to separate dir - HTTP server moved - Files updated to read from new namespaces
