+++ server/static/scripts.js
